TW-1650: reorganize image_builder folder
