+++ lib/pages/chat/sticker_picker_dialog.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/l10n.dart';
 import 'package:matrix/matrix.dart';
 
-import 'events/image_bubble.dart';
+import 'events/images_builder/image_bubble.dart';
 
 class StickerPickerDialog extends StatefulWidget {
   final Room room;
